[SYCLops][TestCoverage] Added DNN Kernel Tests
Added four common kernels used in DNNs (avgpool, convolution, matmul,
and maxpool) based on kernels found in ResNet-50.

These testcases can be used to demonstrate SYCLops lowering kernels
to MLIR as part of a DNN.

Added sycl::max (float) to the list of supported sycl functions and
added the code to handle it in MLIRCodeGen.